refactor: tests runner
