+++ pkg/runner/runner_test.go
@@ -4,21 +4,25 @@ import (
 	"context"
 	"errors"
 	"flag"
-	"time"
 
+	"github.com/kyverno/chainsaw/pkg/apis"
+	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha2"
+	"github.com/kyverno/chainsaw/pkg/client"
+	fake "github.com/kyverno/chainsaw/pkg/client/testing"
 	"github.com/kyverno/chainsaw/pkg/discovery"
 	"github.com/kyverno/chainsaw/pkg/loaders/config"
 	"github.com/kyverno/chainsaw/pkg/model"
 	enginecontext "github.com/kyverno/chainsaw/pkg/runner/context"
 	"github.com/kyverno/chainsaw/pkg/runner/flags"
 	"github.com/kyverno/chainsaw/pkg/runner/internal"
+	"github.com/kyverno/chainsaw/pkg/runner/mocks"
 	"github.com/kyverno/chainsaw/pkg/testing"
 	"github.com/stretchr/testify/assert"
+	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/rest"
 	"k8s.io/utils/clock"
-	tclock "k8s.io/utils/clock/testing"
+	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func TestNew(t *testing.T) {
@@ -52,8 +56,46 @@ func TestNew(t *testing.T) {
 func Test_runner_Run(t *testing.T) {
 	config, err := config.DefaultConfiguration()
 	assert.NoError(t, err)
-	tc, err := InitContext(config.Spec, nil, nil)
-	assert.NoError(t, err)
+	tc := func() enginecontext.TestContext {
+		tc, err := InitContext(config.Spec, nil, nil)
+		assert.NoError(t, err)
+		return tc
+	}
+	echoTest := discovery.Test{
+		Test: &model.Test{
+			TypeMeta: metav1.TypeMeta{
+				APIVersion: "chainsaw.kyverno.io/v1alpha1",
+				Kind:       "Test",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "test",
+			},
+			Spec: v1alpha1.TestSpec{
+				Steps: []v1alpha1.TestStep{{
+					TestStepSpec: v1alpha1.TestStepSpec{
+						Try: []v1alpha1.Operation{{
+							Script: &v1alpha1.Script{
+								Content: "echo hello",
+							},
+						}},
+					},
+				}},
+			},
+		},
+	}
+	mockTC := func(client client.Client) enginecontext.TestContext {
+		registry := mocks.Registry{
+			Client: client,
+		}
+		return enginecontext.MakeContext(apis.NewBindings(), registry).WithTimeouts(v1alpha1.Timeouts{
+			Apply:   &config.Spec.Timeouts.Apply,
+			Assert:  &config.Spec.Timeouts.Assert,
+			Cleanup: &config.Spec.Timeouts.Cleanup,
+			Delete:  &config.Spec.Timeouts.Delete,
+			Error:   &config.Spec.Timeouts.Error,
+			Exec:    &config.Spec.Timeouts.Exec,
+		})
+	}
 	type summaryResult struct {
 		passed  int32
 		failed  int32
@@ -67,23 +109,107 @@ func Test_runner_Run(t *testing.T) {
 		want    *summaryResult
 		wantErr bool
 	}{{
-		name:    "no tests",
-		config:  config.Spec,
-		tc:      tc,
-		tests:   nil,
-		want:    nil,
-		wantErr: false,
+		name:   "no tests",
+		config: config.Spec,
+		tc:     tc(),
+		tests:  nil,
+		want:   nil,
 	}, {
 		name:   "test with err",
 		config: config.Spec,
-		tc:     tc,
+		tc:     tc(),
 		tests:  []discovery.Test{{Err: errors.New("dummy")}},
 		want: &summaryResult{
-			passed:  0,
-			failed:  1,
-			skipped: 0,
+			failed: 1,
 		},
-		wantErr: false,
+	}, {
+		name: "Namesapce exists - success",
+		config: model.Configuration{
+			Namespace: v1alpha2.NamespaceOptions{
+				Name: "default",
+			},
+		},
+		tc: func() enginecontext.TestContext {
+			client := &fake.FakeClient{
+				GetFn: func(ctx context.Context, call int, key ctrlclient.ObjectKey, obj ctrlclient.Object, opts ...ctrlclient.GetOption) error {
+					return nil
+				},
+			}
+			return mockTC(client)
+		}(),
+		tests: []discovery.Test{echoTest},
+		want: &summaryResult{
+			passed: 1,
+		},
+	}, {
+		name: "Namesapce doesn't exists - success",
+		config: model.Configuration{
+			Namespace: v1alpha2.NamespaceOptions{
+				Name: "chain-saw",
+			},
+		},
+		tc: func() enginecontext.TestContext {
+			client := &fake.FakeClient{
+				GetFn: func(ctx context.Context, call int, key ctrlclient.ObjectKey, obj ctrlclient.Object, opts ...ctrlclient.GetOption) error {
+					return kerrors.NewNotFound(v1alpha1.Resource("Namespace"), "chain-saw")
+				},
+				CreateFn: func(ctx context.Context, call int, obj ctrlclient.Object, opts ...ctrlclient.CreateOption) error {
+					return nil
+				},
+				DeleteFn: func(ctx context.Context, call int, obj ctrlclient.Object, opts ...ctrlclient.DeleteOption) error {
+					return nil
+				},
+			}
+			return mockTC(client)
+		}(),
+		tests: []discovery.Test{echoTest},
+		want: &summaryResult{
+			passed: 1,
+		},
+	}, {
+		name: "Namesapce not found - error",
+		config: model.Configuration{
+			Namespace: v1alpha2.NamespaceOptions{
+				Name: "chain-saw",
+			},
+		},
+		tc: func() enginecontext.TestContext {
+			client := &fake.FakeClient{
+				GetFn: func(ctx context.Context, call int, key ctrlclient.ObjectKey, obj ctrlclient.Object, opts ...ctrlclient.GetOption) error {
+					return kerrors.NewBadRequest("failed to get namespace")
+				},
+				CreateFn: func(ctx context.Context, call int, obj ctrlclient.Object, opts ...ctrlclient.CreateOption) error {
+					return nil
+				},
+			}
+			return mockTC(client)
+		}(),
+		tests: []discovery.Test{echoTest},
+		want: &summaryResult{
+			failed: 1,
+		},
+	}, {
+		name: "Namesapce doesn't exists and can't be created - error",
+		config: model.Configuration{
+			Namespace: v1alpha2.NamespaceOptions{
+				Name: "chain-saw",
+			},
+		},
+		tc: func() enginecontext.TestContext {
+			client := &fake.FakeClient{
+				GetFn: func(ctx context.Context, call int, key ctrlclient.ObjectKey, obj ctrlclient.Object, opts ...ctrlclient.GetOption) error {
+					return kerrors.NewNotFound(v1alpha1.Resource("Namespace"), "chain-saw")
+				},
+				CreateFn: func(ctx context.Context, call int, obj ctrlclient.Object, opts ...ctrlclient.CreateOption) error {
+					return kerrors.NewBadRequest("failed to create namespace")
+				},
+			}
+			return mockTC(client)
+		}(),
+		tests: []discovery.Test{echoTest},
+		want: &summaryResult{
+			failed: 1,
+		},
 	}}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -94,16 +220,16 @@ func Test_runner_Run(t *testing.T) {
 			// to allow unit tests to work
 			assert.NoError(t, flags.SetupFlags(tt.config))
 			assert.NoError(t, flag.Set("test.testlogfile", ""))
-			err := r.Run(context.TODO(), tt.config, tt.tc, tt.tests...)
+			err := r.Run(context.TODO(), tt.config.Namespace, tt.tc, tt.tests...)
 			if tt.wantErr {
 				assert.Error(t, err)
 			} else {
 				assert.NoError(t, err)
 			}
 			if tt.want != nil {
-				assert.Equal(t, tt.want.failed, tc.Failed())
-				assert.Equal(t, tt.want.passed, tc.Passed())
-				assert.Equal(t, tt.want.skipped, tc.Skipped())
+				assert.Equal(t, tt.want.failed, tt.tc.Failed())
+				assert.Equal(t, tt.want.passed, tt.tc.Passed())
+				assert.Equal(t, tt.want.skipped, tt.tc.Skipped())
 			}
 		})
 	}
@@ -155,7 +281,7 @@ func Test_runner_run(t *testing.T) {
 			r := &runner{
 				clock: clock.RealClock{},
 			}
-			err := r.run(context.TODO(), tt.m, model.Configuration{}, enginecontext.EmptyContext(), tt.tests...)
+			err := r.run(context.TODO(), tt.m, v1alpha2.NamespaceOptions{}, enginecontext.EmptyContext(), tt.tests...)
 			if tt.wantErr {
 				assert.Error(t, err)
 			} else {
@@ -182,98 +308,5 @@ func Test_runner_onFail(t *testing.T) {
 		}
 		r.onFail()
 		assert.True(t, called)
-	}
-}
-
-func TestRun(t *testing.T) {
-	fakeClock := tclock.NewFakePassiveClock(time.Now())
-	config, err := config.DefaultConfiguration()
-	if err != nil {
-		assert.NoError(t, err)
-	}
-	tests := []struct {
-		name       string
-		tests      []discovery.Test
-		config     model.Configuration
-		restConfig *rest.Config
-		mockReturn int
-		wantErr    bool
-	}{{
-		name:  "Zero Tests",
-		tests: []discovery.Test{},
-		config: model.Configuration{
-			Timeouts: config.Spec.Timeouts,
-		},
-		restConfig: &rest.Config{},
-		wantErr:    false,
-	}, {
-		name: "Nil Rest Config with 1 Test",
-		tests: []discovery.Test{
-			{
-				Err: nil,
-				Test: &model.Test{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "test1",
-					},
-				},
-			},
-		},
-		config: model.Configuration{
-			Timeouts: config.Spec.Timeouts,
-			Report: &v1alpha2.ReportOptions{
-				Format: v1alpha2.JSONFormat,
-			},
-		},
-		restConfig: nil,
-		wantErr:    false,
-	}, {
-		name: "Success Case with 1 Test",
-		tests: []discovery.Test{
-			{
-				Err: nil,
-				Test: &model.Test{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "test1",
-					},
-				},
-			},
-		},
-		restConfig: &rest.Config{},
-		mockReturn: 0,
-		wantErr:    false,
-	}, {
-		name: "Failure Case with 1 Test",
-		tests: []discovery.Test{
-			{
-				Err: nil,
-				Test: &model.Test{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "test1",
-					},
-				},
-			},
-		},
-		restConfig: &rest.Config{},
-		mockReturn: 2,
-		wantErr:    true,
-	}}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockMainStart := &_mainStart{
-				code: tt.mockReturn,
-			}
-			runner := runner{
-				clock: fakeClock,
-			}
-			ctx := context.TODO()
-			tc, err := InitContext(tt.config, tt.restConfig, nil)
-			assert.NoError(t, err)
-			err = runner.run(ctx, mockMainStart, tt.config, tc, tt.tests...)
-			if tt.wantErr {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-		})
 	}
 }
